feat: rename setting card
