--- formular/src/app/pages/setting/setting-card/setting-card.component.html
@@ -1,33 +1,0 @@
-<p-accordion (onOpen)="onTableOpen($event)">
-    <p-accordionTab header="Azure Serive Principle Setting" class="setting-card">
-       <div style="padding-top: 1%;">
-            <p-table #vmTable [value]="servicePrincipleSettings" responsiveLayout="scroll" [globalFilterFields]="['clientId','subscriptionId', 'talentId']">
-                <ng-template pTemplate="header"> 
-                    <tr>
-                        <th pSortableColumn="clientId">ClentId <p-sortIcon field="clientId"></p-sortIcon></th>
-                        <th pSortableColumn="subscriptionId">SubscriptionId<p-sortIcon field="subscriptionId"></p-sortIcon></th>
-                        <th pSortableColumn="talentId">TalentId<p-sortIcon field="talentId"></p-sortIcon></th>
-                        <th pSortableColumn="update">Update</th>
-                        <th pSortableColumn="remove">Remove</th>
-                    </tr>
-                </ng-template>
-                <ng-template pTemplate="body" let-servicePrincipleSettings >
-                    <tr>
-                        <td>{{ servicePrincipleSettings.clientId }}</td>
-                        <td>{{ servicePrincipleSettings.subscriptionId }}</td>
-                        <td>{{ servicePrincipleSettings.talentId }}</td>
-                        <td class="td-button" type="button" (click)="clickUpdate"> 
-                            <i class="pi pi-pencil" style="margin-right: 10%;" ></i>Update
-                        </td>
-                        <td class="td-button" type="button" (click)="clickRemove"> 
-                            <i class="pi pi-trash" style="margin-right: 10%;" ></i>Remove
-                        </td>
-                    </tr>
-                </ng-template>
-            </p-table>
-       </div>
-    </p-accordionTab>
-    <p-accordionTab header="Jenkins Server Setting" class="setting-card">
-        Content 2
-    </p-accordionTab>
-</p-accordion>
