+++ formular/src/app/pages/setting/setting.component.html
@@ -1,3 +1,10 @@
 <div style="padding-top: 1%;">
-    <app-setting-card></app-setting-card>
+    <p-accordion (onOpen)="onTableOpen($event)">
+        <p-accordionTab header="Azure Serive Principle Setting" class="setting-card">
+            <app-azure-sp-setting-card></app-azure-sp-setting-card>
+        </p-accordionTab>
+        <p-accordionTab header="Jenkins Server Setting" class="setting-card">
+            Content 2
+        </p-accordionTab>
+    </p-accordion>
 </div>
